refactor: prune legacy renderer and consolidate tile utilities (A-12)
